+++ routes/statistics.js
@@ -19,7 +19,7 @@ router.get('/:from/:to', function (req, res, next) {
 /** Get statistics fro the past two weeks for all players */
 router.get('/weekly', function (req, res, next) {
     var from = moment().isoWeekday(1).format('YYYY-MM-DD');
-    var to = moment().isoWeekday(7).format('YYYY-MM-DD');
+    var to = moment().isoWeekday(8).format('YYYY-MM-DD');
     getStatistics(from, to, req, res, next);
 });
 
@@ -27,7 +27,7 @@ function getStatistics(from, to, req, res, next) {
     axios.all([
         axios.get(req.app.locals.kcapp.api + '/player'),
         axios.get(req.app.locals.kcapp.api + '/office'),
-        axios.get(req.app.locals.kcapp.api + '/statistics/x01/' + from + '/' + to),
+        axios.get(req.app.locals.kcapp.api + '/statistics/1/' + from + '/' + to),
         axios.get(req.app.locals.kcapp.api + '/statistics/office/' + from + '/' + to)
     ]).then(axios.spread((players, offices, x01, office) => {
         res.marko(statisticsTemplate, {